Added program to calculate the most frequently used words in /r/hockey
Justin Tomlinson

This program will find the most common words on /r/hockey. I will manually
go through this to find the most popular players that I can use in my
sentiment analysis that I will work on later. More information is available
in the README to understand what to do and what it is doing.